requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
PyJWT==2.10.1
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
